backend/requirements_flexible.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-socketio>=5.11.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
celery>=5.3.0
redis>=5.0.0
twilio>=9.0.0
reportlab>=4.0.0
opencv-python-headless>=4.9.0
gTTS>=2.5.0
ultralytics>=8.2.0
torch>=2.6.0
torchvision>=0.21.0
sentence-transformers>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
slowapi>=0.1.9
aiofiles>=23.0.0
